refactor: core:ui 모듈 선언 후 추가 설정 (#38)
- UI 에서만 빈번하게 쓰이는 함수, 클래스 파일 이동
- string resource wrapper class: UiString
- textField 텍스트 검증: validateInput
