+++ core/ui/build.gradle.kts
@@ -1,0 +1,10 @@
+import com.yapp.breake.setNamespace
+
+plugins {
+	alias(libs.plugins.breake.android.library)
+	alias(libs.plugins.breake.android.compose)
+}
+
+android {
+	setNamespace("core.ui")
+}
