+++ presentation/home/src/main/java/com/yapp/breake/presentation/registry/model/RegistrySnackBarState.kt
@@ -1,7 +1,7 @@
 package com.yapp.breake.presentation.registry.model
 
 import androidx.compose.runtime.Immutable
-import com.yapp.breake.core.util.UiString
+import com.yapp.breake.core.ui.UiString
 
 sealed interface RegistrySnackBarState {
 	@Immutable
