+++ presentation/setting/src/main/java/com/yapp/breake/presentation/nickname/component/SettingNicknameTextFeld.kt
@@ -18,7 +18,7 @@ import com.yapp.breake.core.designsystem.theme.Gray200
 import com.yapp.breake.core.designsystem.theme.Green
 import com.yapp.breake.core.designsystem.theme.Red
 import com.yapp.breake.core.designsystem.theme.White
-import com.yapp.breake.core.util.isValidInput
+import com.yapp.breake.core.ui.isValidInput
 
 @Composable
 internal fun SettingNicknameTextField(
