+++ presentation/signup/src/main/java/com/yapp/breake/presentation/signup/component/NicknameTextField.kt
@@ -13,7 +13,7 @@ import com.yapp.breake.core.designsystem.theme.BrakeTheme
 import com.yapp.breake.core.designsystem.theme.Green
 import com.yapp.breake.core.designsystem.theme.Red
 import com.yapp.breake.core.designsystem.theme.White
-import com.yapp.breake.core.util.isValidInput
+import com.yapp.breake.core.ui.isValidInput
 
 @Composable
 fun NicknameTextField(
